% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_t_beam_centroid.R
\name{calculate_t_beam_centroid}
\alias{calculate_t_beam_centroid}
\title{Get centroid t-beam}
\usage{
calculate_t_beam_centroid(.flange, .flange_thickness, .web, .web_thickness)
}
\arguments{
\item{.flange}{Dimension of the flange \link{mm}}

\item{.flange_thickness}{Dimension of the flange thickness \link{mm}}

\item{.web}{Dimension of the web \link{mm}}

\item{.web_thickness}{Dimension of the web thickness \link{mm}}
}
\value{
A data frame with the X-Y coordinates of the centroid
}
\description{
Calculate the centroid of a T-beam (from the bottom)
}
\examples{
calculate_t_beam_centroid(.flange = 38.1,
                          .flange_thickness = 6.4,
                          .web = 38.1,
                          .web_thickness = 6.4)
}
