% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_t_beam.R
\name{gg_t_beam}
\alias{gg_t_beam}
\title{Plot cross-section area}
\usage{
gg_t_beam(.flange, .flange_thickness, .web, .web_thickness, ...)
}
\arguments{
\item{.flange}{Dimension of the flange \link{mm}}

\item{.flange_thickness}{Dimension of the flange thickness \link{mm}}

\item{.web}{Dimension of the web \link{mm}}

\item{.web_thickness}{Dimension of the web thickness \link{mm}}
}
\value{
A ggplot2 object
}
\description{
Plot in ggplot2 the cross-sectional area of a T-beam
}
\examples{
library(ggplot2)

centroid <- calculate_t_beam_centroid(.flange = 38.1,
                                      .flange_thickness = 6.4,
                                      .web = 38.1,
                                      .web_thickness = 6.4)

I <- calculate_t_beam_I_centroid(.flange = 38.1,
                                 .flange_thickness = 6.4,
                                 .web = 38.1,
                                 .web_thickness = 6.4)

gg_t_beam(.flange = 38.1,
          .flange_thickness = 6.4,
          .web = 38.1,
          .web_thickness = 6.4,
          fill = "darkgrey") +
  geom_point(data = centroid, aes(x, y), col = "red", size = 4) +
  geom_text(data = centroid, aes(x = x,
                                 y = y,
                                 label = paste0("Centroid (", round(x, 2), ", ", round(y, 2), ")"),
                                 hjust = -0.1,
                                 vjust = 0),
            size = 6,
            color = "red") +
  annotate(geom = "text", x = 5.6, y = 40, label = paste("Ix = " , round(I[1]), "mm4"), hjust = "left") +
  annotate(geom = "text", x = 5.6, y = 38, label = paste("Iy = " , round(I[2]), "mm4"), hjust = "left") +
  theme_bw()
}
