% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_internal_forces.R
\name{plot_internal_forces}
\alias{plot_internal_forces}
\title{Shear and moment diagrams}
\usage{
plot_internal_forces(.beam_length, .loads, .resolution = 0.01)
}
\arguments{
\item{.beam_length}{Longitude of the beam in meters}

\item{.loads}{A tibble with punctual forces, it must have two columns (x and force).
\code{x} is the position of the punctual force on the beam from the left edge.
\code{force} is the magnitude of the force; positive forces going up, and negative
forces going down. Support reactions should also be added in this data frame.}

\item{.resolution}{The beam is discretised for the calculations, by default
by steps of 0.01 m (but it can be adjusted if necessary).}
}
\value{
A ggplot2 object with the loads, shear diagram, and the bending moment.
}
\description{
Plot the shear and bending moment diagrams of a beam
}
\examples{
library(ggplot2)
library(patchwork)

loads <- tibble::tribble( ~x, ~force,
                          0.025,  50,
                          0.325, -50,
                          0.500, -50,
                          0.800,  50)

plot_internal_forces(.beam_length = 0.825, .resolution = 0.0001, .loads = loads)
}
